requirements.txt
aiogram
aiohttp
aiosignal
async-timeout
attrs
Babel
beautifulsoup4
bs4
certifi
charset-normalizer
frozenlist
idna
lxml
magic-filter
multidict
psycopg2-binary
python-dotenv
pytz
requests
soupsieve
urllib3
yarl
